Feature/about and profile routes

+++ controllers/api/usersRoutes.js
@@ -4,30 +4,24 @@ const { User } = require('../../models');
 router.post('/login', async (req, res) => {
   try {
     const userData = await User.findOne({ where: { email: req.body.email } });
-
     if (!userData) {
       res
         .status(400)
         .json({ message: 'Incorrect email or password, please try again' });
       return;
     }
-
     const validPassword = await userData.checkPassword(req.body.password);
-
     if (!validPassword) {
       res
         .status(400)
         .json({ message: 'Incorrect email or password, please try again' });
       return;
     }
-
     req.session.save(() => {
       req.session.user_id = userData.id;
       req.session.logged_in = true;
-      
       res.json({ user: userData, message: 'You are now logged in!' });
     });
-
   } catch (err) {
     res.status(400).json(err);
   }
@@ -36,36 +30,38 @@ router.post('/login', async (req, res) => {
 router.post('/signup', async (req, res )=> {
     try {
         const userData = await User.create(req.body);
-
     if (!userData) {
       res
         .status(400)
         .json({ message: 'Incorrect email or password, please try again' });
       return;
     }
-
     req.session.save(() => {
       req.session.user_id = userData.id;
       req.session.logged_in = true;
-      
       res.json({ user: userData, message: 'You are now logged in!' });
     })
-
       } catch (err) {
         res.status(400).json(err);
-      }
+      };
 })
 
+// Delete Route to remove a User.
 router.delete('/:user_id', (req, res) =>{
-  User.destroy({
-    where: {
-      user_id: req.params.user_id
-    }
-  })
-  .then((deletedUser) => {
-    res.json(deletedUser);
-  })
-  .catch((err) => res.json(err))
+  if(req.params.user_id === req.session.user_id) {
+   User.destroy({
+     where: {
+       user_id: req.params.user_id
+     }
+   })
+   .then((deletedUser) => {
+     res.json(deletedUser);
+   })
+   .catch((err) => res.json(err));
+   }
+   else{
+    res.status(400).json("Error! Can not delete another user!");
+   };
 })
 
 router.post('/logout', (req, res) => {
